Separate partners data into a separate file from the helpers one
